continuous zoom: the zoom level becomes a real-valued double resting anywhere in [1,10] (fit stays the 0.0 sentinel), spp derives from 0.625·2^(level−1) ms/px, tab_X_zoom persists as a min-0 round-trip double in both binaries, and c jumps the playhead to the focused marker before snapping to level 2

+++ src/gui/render_pipeline.h
@@ -65,7 +65,7 @@ struct AuthoringSnapshot {
     // clamps them into the entry's own map domain before writing (a sweep
     // cell's rewritten markers can give the cell a shorter target axis).
     int64_t view_viewport_start_frame = 0;   // target-axis frames
-    int     view_zoom_level           = 0;   // persisted zoom vocabulary
+    double  view_zoom_level           = 0.0; // persisted zoom vocabulary
     int64_t view_playhead_frame       = 0;   // target-axis frames
 };
